+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.2] - 2022-12-02
+
 ### Fixed 
 
 - Cleanup all annotations.
@@ -50,7 +52,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Init
 - Docs
 
-[Unreleased]: https://github.com/giantswarm/aws-rolling-node-operator/compare/v0.3.1...HEAD
+[Unreleased]: https://github.com/giantswarm/aws-rolling-node-operator/compare/v0.3.2...HEAD
+[0.3.2]: https://github.com/giantswarm/aws-rolling-node-operator/compare/v0.3.1...v0.3.2
 [0.3.1]: https://github.com/giantswarm/aws-rolling-node-operator/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/giantswarm/aws-rolling-node-operator/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/giantswarm/aws-rolling-node-operator/compare/v0.1.2...v0.2.0